requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.23
aiohttp>=3.8.1
pandas>=1.3.3
numpy>=1.21.2
scikit-learn>=0.24.2
python-telegram-bot>=13.7
lightgbm>=3.3.2
optuna>=2.10.0
shap>=0.40.0
python-jose>=3.3.0
python-multipart>=0.0.5
ta-lib>=0.4.24
plotly>=5.3.1
dash>=2.0.0
reportlab>=3.6.2
cryptography>=3.4.7
pytest>=6.2.5
black>=21.9b0
flake8>=3.9.2
mypy>=0.910
pytest-asyncio>=0.15.1
